requirements.txt
numpy
amulet-core
opencv-python 
opencv-contrib-python
Pillow